Rename EdgeScrollContainer to Carousel
Rename the component and file to a standard, recognizable name:
- EdgeScrollContainer → Carousel
- EdgeScrollItem → CarouselItem
- edge-scroll-container.tsx → carousel.tsx
- Embla carousel → embla-carousel.tsx (used only by ReleaseCarousel)
