+++ app/[locale]/developers/tools/page.tsx
@@ -9,10 +9,7 @@ import MainArticle from "@/components/MainArticle"
 import SubpageCard from "@/components/SubpageCard"
 import { Button } from "@/components/ui/buttons/Button"
 import { Card } from "@/components/ui/card"
-import {
-  EdgeScrollContainer,
-  EdgeScrollItem,
-} from "@/components/ui/edge-scroll-container"
+import { Carousel, CarouselItem } from "@/components/ui/carousel"
 import { LinkBox, LinkOverlay } from "@/components/ui/link-box"
 import { Section } from "@/components/ui/section"
 
@@ -87,9 +84,9 @@ const Page = async (props: {
 
         <Section id="apps" className="space-y-4">
           <h2>{t("page-developers-tools-applications-title")}</h2>
-          <EdgeScrollContainer>
+          <Carousel>
             {DEV_TOOL_CATEGORIES.map(({ slug, Icon }) => (
-              <EdgeScrollItem
+              <CarouselItem
                 key={slug}
                 className="ms-6 w-[calc(100%-4rem)] max-w-md md:min-w-96 md:flex-1 lg:max-w-[33%]"
               >
@@ -133,9 +130,9 @@ const Page = async (props: {
                     />
                   ))}
                 </Card>
-              </EdgeScrollItem>
+              </CarouselItem>
             ))}
-          </EdgeScrollContainer>
+          </Carousel>
         </Section>
 
         <Section id="categories" className="space-y-4">
